+++ src/main/java/net/sf/jabref/gui/actions/Actions.java
@@ -16,6 +16,7 @@ public class Actions {
     public static final String COPY_KEY = "copyKey";
     public static final String COPY_CITE_KEY = "copyCiteKey";
     public static final String COPY_KEY_AND_TITLE = "copyKeyAndTitle";
+    public static final String COPY_KEY_AND_LINK = "copyKeyAndLink";
     public static final String CUT = "cut";
     public static final String DB_CONNECT = "dbConnect";
     public static final String DELETE = "delete";